style: fix code in generator
