+++ src/generator/index.ts
@@ -1,3 +1,9 @@
-export { GeneratorOptions } from "./types/generatorOptions";
-import FrostGenerator from "./core";
-export { FrostGenerator };
+export interface GeneratorOptions {
+    srcDir: string;
+    buildDir: string;
+    staticDir: string;
+    HTMLcompressionLevel: number;
+    metadataFile: string;
+}
+
+export * as FrostGenerator from "./core";
