--- src/generator/types/generatorOptions.ts
@@ -1,7 +1,0 @@
-export interface GeneratorOptions {
-    srcDir: string;
-    buildDir: string;
-    staticDir: string;
-    HTMLcompressionLevel: number;
-    metadataFile: string;
-}
